requirements.txt
Django==5.1.7
Pillow==11.1.0
asgiref==3.8.1
sqlparse==0.5.3
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-docx>=0.8.11 